Use correct parameter names in web app deploy
It is not clear from the docs, but looking at the source code for the AzureWebAppContainer task it need a resource group and an image name (as opposed to container name). These parameters deploy successfully for me.

I'm not sure about the azureSubscription value, it might not be needed but I've left it as it was.