Fix CD deployment break (#305)
Recently the step to get the model version had a name associated with the step. We need to also update the references in future step to use the step name as a prefix. This resulted in the MODEL_VERSION variable causing failures in each CD deployment step.

This did not show up in CI because the MODEL_VERSION var is hard coded in the variable group. We should also remove that.